backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
email-validator==2.1.0
requests==2.31.0
urllib3==2.0.7
beautifulsoup4==4.12.2
dnspython==2.4.2
python-dotenv==1.0.0
lxml==4.9.3
html5lib==1.1